models/requirements.txt
Flask==1.1.1
future==0.17.1
imutils==0.5.3
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.1
numpy==1.16.5
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
pandas==0.25.1
pickleshare==0.7.5
Pillow==6.2.0
pylint==2.3.1
requests==2.22.0
scipy==1.3.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tqdm==4.36.1
wrapt==1.11.2
xlrd==1.2.0
zipp==0.6.0
